requirements.txt
Twisted==23.8.0
PyYAML==6.0.1
Jinja2==3.1.2
psycopg2-binary==2.9.7
aiopg==1.4.0
pytest==7.4.2
pytest-twisted==1.14.0
pytest-asyncio==0.21.1